requirements.txt
# ✅ Archivo: requirements.txt
# 👉 Versión estable con Supabase 1.0.3 sin conflictos con APIError

# --- Núcleo Flask y extensiones ---
flask==2.2.5
flask_session==0.4.0
flask_socketio==5.3.3
flask-wtf==1.1.1  # Para protección CSRF y manejo de formularios
eventlet==0.33.3

# --- Cliente Supabase y dependencias ---
supabase==1.0.3
httpx==0.23.3
postgrest==0.10.7  # Versión moderna sin APIError
pydantic>=2.1.0,<3.0  # Compatible con postgrest
python-dotenv==1.0.1

# --- Librerías adicionales útiles ---
pytz==2023.3
python-dateutil==2.8.2
requests==2.31.0
openai==1.3.7
apscheduler==3.10.4

# --- Compatibilidad con async/await ---
dnspython==2.3.0

# --- Otros componentes que usas ---
gevent==24.11.1
Cython==0.29.23
cffi>=1.15.1
pdfkit==1.0.0
PyPDF2>=3.0.0
fpdf2==2.8.3
pandas==2.2.2
facebook-business==17.0.0
google-ads==22.0.0
google-auth-oauthlib==1.2.0
qrcode[pil]==8.2
flask-cors==4.0.0
python-socketio==5.8.0
flask-apscheduler==1.12.3

# --- Compatibilidad Flask (evita conflictos con jinja, werkzeug, etc) ---
werkzeug>=2.2.2,<3.0
itsdangerous==2.0.1
jinja2==3.0.3
markupsafe>=2.1.1,<3.0
